+++ src/components/visual/transaction-list-item/transaction-list-item.types.ts
@@ -1,3 +1,0 @@
-export * from '../TransactionListItem/transactionListItem.types';
-
-// This file is a temporary workaround to maintain backward compatibility with sdk-dapp v5.5.1.